Code/1steering/packages/requirements.txt
# Core dependencies for diffusion model training and interpretability

# PyTorch (install separately with CUDA support)
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and diffusion
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Data processing
Pillow>=9.5.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Interactive interface
gradio>=3.35.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development tools (optional)
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
